user edit

+++ pages/api/chapter/[id].js
@@ -1,6 +1,7 @@
 import CourseChapter from "../../../models/course-chapter"
 import { loginCheck } from "../../../utils/firebaseAdmin"
 import { serializeChapter } from "../course/utils"
+import Profile from "../../../models/profile"
 
 export default async function handler(req, res){
   try {
@@ -8,6 +9,10 @@ export default async function handler(req, res){
   } catch (error) {
     return res.status(401).send('FORBIDDEN')
   }
+  await dbConnect()
+  const profile = await Profile.findOne({email: req.decodedClaims.email})
+  if(!profile?.admin)
+    return res.status(401).send('FORBIDDEN')
   if(req.method==='DELETE')
   {
     await CourseChapter.findByIdAndRemove(req.query.id)
